fastapi_backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
psycopg2
pydantic[email]
python-jose
passlib
python-multipart